Further improvements
- Operation Failure is now a data class
- Threading is now explicitly controlled by the part requiring it (don't just shovel things to computation)
