+++ app/src/main/java/com/futurice/freesound/feature/home/HomeFragmentViewModel.kt
@@ -16,15 +16,14 @@
 
 package com.futurice.freesound.feature.home
 
-import com.futurice.freesound.feature.common.streams.Fetch
-import com.futurice.freesound.feature.common.streams.Operation
 import com.futurice.freesound.mvi.*
-import com.futurice.freesound.network.api.model.User
 import timber.log.Timber
 
-internal class HomeFragmentViewModel(private val tag: String, store: Store<HomeUiModel, Action, Result>)
-    : BaseViewModel<UiEvent, Action, Result, HomeUiModel>("HomeFragmentViewModel",
-        initialEvent = UiEvent.InitialEvent,
+internal class HomeFragmentViewModel(tag: String = "HomeFragmentViewModel",
+                                     initialEvent: UiEvent = UiEvent.Initial,
+                                     store: Store<HomeUiModel, Action, Result>)
+    : BaseViewModel<UiEvent, Action, Result, HomeUiModel>(tag = tag,
+        initialEvent = initialEvent,
         store = store) {
 
     companion object {
@@ -38,23 +37,10 @@ internal class HomeFragmentViewModel(private val tag: String, store: Store<HomeU
 
     override fun eventToAction(uiEvent: UiEvent): Action {
         return when (uiEvent) {
-            UiEvent.InitialEvent -> Action.Initial
+            UiEvent.Initial -> Action.Initial
             UiEvent.RefreshRequested -> Action.RefreshContent
             UiEvent.ErrorIndicatorDismissed -> Action.ClearError
         }.also { Timber.d("From UiEvent: $uiEvent, Action is: $it") }
     }
 
-}
-
-sealed class Result(val log: String) {
-    object NoChange : Result("No-op change")
-    object ErrorCleared : Result("Error dismissed change")
-    data class Refreshed(val refresh: Operation) : Result("User Fetch state change: $refresh")
-    data class UserUpdated(val userFetch: Fetch<User>) : Result("User state change: $userFetch")
-}
-
-sealed class Action(val log: String) {
-    object Initial : Action("Initial action")
-    object ClearError : Action("Error cleared action")
-    object RefreshContent : Action("Content refresh action")
 }
