+++ app/src/main/java/com/futurice/freesound/mvi/Store.kt
@@ -1,6 +1,5 @@
 package com.futurice.freesound.mvi
 
-import com.futurice.freesound.feature.common.scheduling.SchedulerProvider
 import io.reactivex.FlowableTransformer
 import timber.log.Timber
 
@@ -8,8 +7,7 @@ class Store<S, A, R>(
         private val tag: String,
         private val initialState: S,
         private val actionTransformer: ActionTransformer<A, R>,
-        private val reducer: Reducer<R, S>,
-        private val schedulers: SchedulerProvider) {
+        private val reducer: Reducer<R, S>) {
 
     fun dispatchAction(): FlowableTransformer<A, S> {
         return FlowableTransformer { it ->
@@ -18,7 +16,6 @@ class Store<S, A, R>(
                     .doOnNext { model: S -> Timber.v(" $model") }
                     .doOnError { e: Throwable -> Timber.e(e, "An unexpected fatal error occurred in $tag") }
                     .asUiModelFlowable()
-                    .subscribeOn(schedulers.computation())
         }
     }
